+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added Typecasting#typecaster_for
 * Changed Typecasting#typecast_attribute to take a typecaster, not a type
 * Removed Typecasting#typecast_value
+* TypecastedAttributes now supports a :typecaster option on attribute
+  definitions which can be any object that responds to #call
 
 # ActiveAttr 0.5.0 (March 11, 2012) #